feat(rpc): remove deployed flag and addContracts
