--- yarn-project/aztec-rpc/src/contract_data_source/index.ts
@@ -1,3 +1,0 @@
-export * from './contract_dao.js';
-export * from './contract_data_source.js';
-export * from './memory_contract_data_source.js';
