+++ yarn-project/aztec-rpc/src/contract_database/contract_dao.ts
@@ -10,7 +10,6 @@ export interface ContractDao extends ContractAbi {
   address: AztecAddress;
   portalAddress: EthAddress;
   functions: ContractFunctionDao[];
-  deployed: boolean;
 }
 
 export function functionAbiToFunctionDao(abi: FunctionAbi) {
@@ -24,12 +23,10 @@ export function contractAbiToContractDao(
   address: AztecAddress,
   portalAddress: EthAddress,
   abi: ContractAbi,
-  deployed: boolean,
 ): ContractDao {
   return {
     address,
     portalAddress,
     functions: abi.functions.map(functionAbiToFunctionDao),
-    deployed,
   };
 }
