+++ yarn-project/aztec-rpc/src/contract_database/contract_database.ts
@@ -4,7 +4,6 @@ import { ContractDao } from './contract_dao.js';
 
 export interface ContractDataSource {
   addContract(address: AztecAddress, portalAddress: EthAddress, abi: ContractAbi, deployed?: boolean): Promise<void>;
-  confirmContractsDeployed(addresses: AztecAddress[]): Promise<void>;
   getContract(address: AztecAddress): Promise<ContractDao | undefined>;
   getCode(contractAddress: AztecAddress, functionSelector: Buffer): Promise<string | undefined>;
 }
