+++ yarn-project/aztec-rpc/src/contract_database/index.ts
@@ -1,0 +1,3 @@
+export * from './contract_dao.js';
+export * from './contract_database.js';
+export * from './memory_contract_database.js';
