+++ yarn-project/aztec-rpc/src/database/database.ts
@@ -1,6 +1,6 @@
 import { TxHash } from '@aztec/tx';
 
-import { ContractDataSource } from '../contract_data_source/index.js';
+import { ContractDataSource } from '../contract_database/index.js';
 import { TxDao } from './tx_dao.js';
 
 export interface Database extends ContractDataSource {
